# src/main/resources/ecommers-swagger-openapi-code-generator.yaml
openapi: 3.0.3
info:
  title: Simple E-commerce API
  description: Backend API for product listing, cart, orders, and user authentication.
  version: 1.0.0
  contact:
    email: sreenivasakurabalakota@gmail.com
    name: Kurabalakota Sreenivasa
servers:
  - url: http://localhost:7896
...
paths:
...
  /user/byemail/{email}:
    get:
      summary: Get an user by email
      tags:
        - Users
      operationId: getUserByEmail
      parameters:
        - in: path
          name: email
          description: email of the user
          required: true
          schema:
            type: string
            format: email
            pattern: "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
      responses:
        '200':
          description: User Fetched Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
  /user/{id}:
    get:
      summary: Get an user by ID
      tags:
        - Users
      operationId: getUserById
      parameters:
        - in: path
          name: id
          description: Id of the user
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: User Fetched Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

    patch:
      summary: update an user by ID
      tags:
        - Users
      operationId: updateUserById
      parameters:
        - in: path
          name: id
          description: ID of the user to update
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '400':
          description: Invalid Request

    delete:
      tags:
        - Users
      operationId: deleteUserById
      summary: Delete an user by ID
      parameters:
        - in: path
          name: id
          description: ID of user to delete
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
  /user/{id}/update:

    patch:
      summary: to update user password
      tags:
        - Users
      operationId: toUpdatePassword
      parameters:
        - in: path
          name: id
          description: User Id
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  type: string
                  pattern: "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
                  minLength: 10
                  maxLength: 14
                newPassword:
                  type: string
                  pattern : "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
                  minLength: 10
                  maxLength: 14
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Incorrect old password
          content:
            application/json:
              schema:
                type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1234
        name:
          type: string
          example: John Wick
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
        email:
          type: string
          example: xyz@gmail.com
        createdDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        updatedAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        phoneNumber:
          type: string
        address:
          type: string
    UserUpdateRequest:

      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 40
        password:
          type: string
          pattern: "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
          minLength: 10
          maxLength: 14
        email:
          type: string
          format: email
          pattern: "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
        phoneNumber:
          type: string
          pattern: "^[0-9]{10}$"
        address:
          type: string
          minLength: 3
          maxLength: 120
...